requirements.txt
fastapi~=0.115.5
uvicorn~=0.32.1
pydantic~=2.10.1
numpy~=2.1.3
pandas~=2.2.3
starlette~=0.41.3
PyYAML~=6.0.2